main/requirements.txt
altgraph==0.16.1
arrow==0.12.1
binaryornot==0.4.4
boto3==1.9.74
botocore==1.12.74
botometer==1.3
briefcase==0.2.8
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colosseum==0.1.6
cookiecutter==1.6.0
docutils==0.14
future==0.17.1
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
Jinja2==2.10
jinja2-time==0.2.0
jmespath==0.9.3
macholib==1.11
MarkupSafe==1.1.0
oauthlib==2.1.0
pefile==2018.8.8
poyo==0.4.2
PyInstaller==3.4
PySocks==1.6.8
python-dateutil==2.7.5
python-twitter==3.5
requests==2.21.0
requests-oauthlib==1.0.0
rprint==0.0.8
rpy2==2.9.5
rubicon-objc==0.3.0
s3transfer==0.1.13
six==1.12.0
smmap2==2.0.5
toga==0.3.0.dev11
toga-cocoa==0.3.0.dev11
toga-core==0.3.0.dev11
travertino==0.1.2
tweepy==3.7.0
urllib3==1.23
voc==0.1.6
whichcraft==0.5.2
